Ref: Move submission disabling to admin page
