+++ frontend/src/App.tsx
@@ -15,7 +15,6 @@ import { Stats } from './components/stats/Stats';
 import { Admin } from './components/admin/Admin';
 
 function App() {
-  const [semesterOver, setSemesterOver] = useState<boolean>(false);
   const [errorMessage, setErrorMessage] = useState<string | null>(null);
   const [isAdmin, setIsAdmin] = useState<boolean>(!!localStorage.getItem('isAdmin'));
   const [impersonating, setImpersonating] = useState<boolean>(!!localStorage.getItem('impersonatedUser'));
@@ -66,8 +65,6 @@ function App() {
           setSubmissions={setSubmissions}
           isAdmin={isAdmin}
           setIsAdmin={setIsAdmin}
-          semesterOver={semesterOver}
-          setSemesterOver={setSemesterOver}
         />
         <Routes>
           <Route path="/admin-login" element={<AdminLogin setErrorMessage={setErrorMessage} />} />
