+++ frontend/src/model/service/AdminService.ts
@@ -1,0 +1,9 @@
+import { ServerFacade } from '../../network/ServerFacade';
+
+export class AdminService {
+  private serverFacade: ServerFacade = new ServerFacade();
+
+  async toggleSemesterOver() {
+    return await this.serverFacade.toggleSemesterOver();
+  }
+}
